Supports setting cell styles

+++ src/main/java/org/javaexcel/model/CellType.java
@@ -9,33 +9,10 @@ package org.javaexcel.model;
  */
 public enum CellType {
     NUMBER,
-    DOUBLE,
+    INT,
     DATE,
     MONEY,
     TEXT,
     PERCENT,
     LIST;
-
-    // private int index;
-    // private String typeName;
-    //
-    // private CellType(int index, String typeName) {
-    //
-    // }
-    //
-    // public int getIndex() {
-    // return index;
-    // }
-    //
-    // public void setIndex(int index) {
-    // this.index = index;
-    // }
-    //
-    // public String getTypeName() {
-    // return typeName;
-    // }
-    //
-    // public void setTypeName(String typeName) {
-    // this.typeName = typeName;
-    // }
 }
